requirements.txt
SQLAlchemy==1.1.4
git+https://github.com/samuelcolvin/arq.git@master#egg=arq
aiohttp==1.2.0
aiopg==0.13.0
python-dateutil==2.6.0
trafaret==0.7.5
trafaret-config==1.0.1
pillow==4.0.0
